+++ components/[pageId]/BoardPageView/BoardPageView.tsx
