--- components/[pageId]/BoardView/index.ts
@@ -1,1 +1,0 @@
-export * from './BoardView';
