+++ components/[pageId]/DocumentPageView/components/PageBanner.tsx
