+++ components/[pageId]/DocumentPageView/index.tsx
@@ -1,0 +1,3 @@
+import PageEditor from './DocumentPageView';
+
+export default PageEditor;
