--- components/[pageId]/DocumentView/index.tsx
@@ -1,3 +1,0 @@
-import PageEditor from './DocumentView';
-
-export default PageEditor;
