+++ components/common/BoardEditor/focalboard/src/components/cardDetail/cardDetail.tsx
@@ -4,7 +4,7 @@ import EmojiEmotionsIcon from '@mui/icons-material/EmojiEmotions'
 import ImageIcon from '@mui/icons-material/Image'
 import { Box } from '@mui/material'
 import { BountyIntegration } from 'components/bounties/BountyIntegration'
-import PageBanner, { PageCoverGalleryImageGroups } from 'components/[pageId]/DocumentView/components/PageBanner'
+import PageBanner, { PageCoverGalleryImageGroups } from 'components/[pageId]/DocumentPageView/components/PageBanner'
 import { randomIntFromInterval } from 'lib/utilities/random'
 import { useCallback, useEffect, useRef, useState } from 'react'
 import { FormattedMessage } from 'react-intl'
